requirements.txt
blinker
cachetools
certifi
charset-normalizer
click
Flask
Flask-Uploads
google-api-core
google-auth
google-cloud-core
google-cloud-storage
google-crc32c
google-resumable-media
googleapis-common-protos
idna
itsdangerous
Jinja2
MarkupSafe
proto-plus
protobuf
psycopg2-binary
pyasn1
pyasn1_modules
python-dotenv
requests
rsa
urllib3
Werkzeug
Pytest